Remove unnecessary boxing/unboxing.

+++ src/magic/generator/Ability_Mono_DeckGenerator.java
@@ -78,7 +78,7 @@ public class Ability_Mono_DeckGenerator extends RandomDeckGenerator {
         final ArrayList<String> a = new ArrayList<String>();
 
         for (final MagicColor c : countColors.keySet()) {
-            if (countColors.get(c).intValue() > MIN_NUM_CARDS_WITH_SUBTYPE) {
+            if (countColors.get(c) > MIN_NUM_CARDS_WITH_SUBTYPE) {
                 a.add("" + c.getSymbol());
             }
         }
